fix: share responsive CSS orchestration
